Switch to a pared down arduino-pico framework
Switched to a custom branch of the arduino-pico framework that
has the proper compiled static libraries with the compile flags
changed to be compatible with the SCSI DaynaPORT emulator on
then ZuluSCSI Pico using a Pico W.

The Pico W on board LED blinks instead of the carrier board
LED in these cases:
 1) One long blink when the WiFi interface is initialized
 2) One short blink when a WiFi interface attempts to join a network
 3) Three short blinks when a network has successfully joined
