+++ platformio.ini
@@ -123,10 +123,12 @@ build_flags =
     -DUSE_ARDUINO=1
     -DZULUSCSI_PICO
 
+; Build for the ZuluSCSI Pico carrier board with a Pico-W
+; for SCSI DaynaPORT emulation
 [env:ZuluSCSI_Pico_Net_DaynaPORT]
 platform = https://github.com/maxgerhardt/platform-raspberrypi.git
 platform_packages = platformio/toolchain-gccarmnoneeabi@1.100301.220327
-    framework-arduinopico@https://github.com/BlueSCSI/arduino-pico-internal.git#e139b9c7816602597f473b3231032cca5d71a48a
+    framework-arduinopico@https://github.com/rabbitholecomputing/arduino-pico.git#v3.6.0-DaynaPORT
 framework = arduino
 board = rpipicow
 board_build.core = earlephilhower
@@ -148,7 +150,10 @@ build_flags =
     -DZULUSCSI_PICO
     -DZULUSCSI_NETWORK
     -DZULUSCSI_DAYNAPORT
-
+; build flags mirroring the framework-arduinopico#v3.6.0-DaynaPORT static library build
+    -DPICO_CYW43_ARCH_POLL=1
+	-DCYW43_LWIP=0
+	-DCYW43_USE_OTP_MAC=0
 
 ; Variant of RP2040 platform, based on Raspberry Pico board and a carrier PCB
 ; Differs in pinout from ZuluSCSI_RP2040 platform, but shares most of the code.
